requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
pypdf
python-dotenv
streamlit
